chore: rename sdl stuff to `sdl_*.c`
